Moved table creation to its own file.
