+++ db/tables/ddl/__init__.py
@@ -1,0 +1,1 @@
+from db.tables.ddl.create import *  # noqa
