+++ db/tables/operations.py
@@ -2,7 +2,7 @@
 This file contains functions that involve altering tables at the database level.
 """
 
-from sqlalchemy import Column, String, Table, MetaData, func, select, ForeignKey, literal, exists
+from sqlalchemy import Column, func, select, ForeignKey, literal, exists
 from sqlalchemy.schema import DDLElement, DropTable
 from sqlalchemy.ext import compiler
 from alembic.migration import MigrationContext
@@ -10,44 +10,12 @@ from alembic.operations import Operations
 from sqlalchemy.exc import NoSuchTableError, InternalError
 from psycopg2.errors import DependentObjectsStillExist
 
-from db import columns, constants, schemas
+from db import columns, constants
+from db.tables.ddl.create import create_mathesar_table
 from db.tables.utils import reflect_table
 
 
 SUPPORTED_TABLE_UPDATE_ARGS = {'name', 'sa_columns'}
-
-
-def create_mathesar_table(name, schema, columns_, engine, metadata=None):
-    """
-    This method creates a Postgres table in the specified schema using the
-    given name and column list.  It adds internal mathesar columns to the
-    table.
-    """
-    columns_ = columns.init_mathesar_table_column_list_with_defaults(columns_)
-    schemas.create_schema(schema, engine)
-    # We need this so that we can create multiple mathesar tables in the
-    # same MetaData, enabling them to reference each other in the
-    # SQLAlchemy context (e.g., for creating a ForeignKey relationship)
-    if metadata is None:
-        metadata = MetaData(bind=engine, schema=schema)
-    table = Table(
-        name,
-        metadata,
-        *columns_,
-        schema=schema
-    )
-    table.create(engine)
-    return table
-
-
-def create_string_column_table(name, schema, column_names, engine):
-    """
-    This method creates a Postgres table in the specified schema, with all
-    columns being String type.
-    """
-    columns_ = [Column(column_name, String) for column_name in column_names]
-    table = create_mathesar_table(name, schema, columns_, engine)
-    return table
 
 
 class DropTableCascade(DropTable):
