+++ db/tests/test_columns.py
@@ -10,6 +10,7 @@ from sqlalchemy import (
 )
 from sqlalchemy.exc import IntegrityError
 from db import columns, constants, constraints
+from db.tables import ddl as table_ddl
 from db.tables import operations as table_operations
 from db.tables import utils as table_utils
 from db.types import email, alteration
@@ -285,7 +286,7 @@ def test_rename_column_foreign_keys(engine_with_schema):
     engine, schema = engine_with_schema
     table_name = "table_to_split"
     columns_list = [Column("Filler 1", Integer), Column("Filler 2", Integer)]
-    table_operations.create_mathesar_table(table_name, schema, columns_list, engine)
+    table_ddl.create_mathesar_table(table_name, schema, columns_list, engine)
     extracted, remainder, fk_name = table_operations.extract_columns_from_table(
         table_name, ["Filler 1"], "Extracted", "Remainder", schema, engine
     )
@@ -302,7 +303,7 @@ def test_rename_column_sequence(engine_with_schema):
     new_col_name = "new_" + constants.ID
     engine, schema = engine_with_schema
     table_name = "table_with_columns"
-    table = table_operations.create_mathesar_table(table_name, schema, [], engine)
+    table = table_ddl.create_mathesar_table(table_name, schema, [], engine)
     with engine.begin() as conn:
         ins = table.insert()
         conn.execute(ins)
